Apache https configuration and key generation
A number of changes to the original apache-sites-enabled.md file, as the
example domain used originally, "wiki.com" exists. Wanted to use a
non-existant domain for our examples so that it doesn't step on any real
web sites toes. In addition, added the section on using https, rather
than http, as nearly every web site today should be running with a
certificate. Also added a separate procedure, ssl_keys_https.md which
deals with the process of generating the private keys, certificate
signing request and general instructions on using the CSR for purchasing
a certificate.