Vsftpd should be a stand-alone document
Per discussion with Wale, the _vsftpd_ server should be stand-alone. It
isn't necessarily required to only be installed on a web server. To add
clarity, deleted the old named file
(apache_hardened_webserver_vsftpd.md) and saved the file as
secure_ftp_server_vsftpd.md. This required editing the
apache_hardened_webserver.md main document to change the link to the ftp
server process.

The secure_ftp_server_vsftpd.md also reworded the document so that it
did not specify use on a web server, but also included using it on a web
server as an option.