Update README.md
There were many grammatical errors and at least one spelling error. I have written a great amount of documentation and have also done a good amount of technical writing. I also know that it is sometimes difficult to proofread one's own writing. I can substantiate the reasons for each of these changes (in particular the contractions do not belong in technical writing or documentation). We also do not have big goals ... Goals do not have sizes; they may be great, insignificant, or minor but the cannot be quantified. I am seriously interested in promoting 'Rocky Linux' and our main focus should be on a professional and understandable face to the documentation and of course good code ... which I intend to contrubute to also..