Don't recommend the -y switch to DNF.
`dnf -y` is great for scripting, but when recommended to an end-user and that
user uses the -y switch it takes away the user's ability to examine what a
particular dnf command will will and stop it if it's not what he intended.  For
that reason it is best not to recommend to people to use `-y` with dnf and
simply let them accept the changes interactively.