Clarify cron-how-to document
* This edit was done post lab tests uninstalling anacron and installing
noanacron. It changed a bunch of the references in this document because
of the outcomes of the test.
* Tweaked the title: meta
* made sure commands were in code blocks
* modified the cron "dot" directories section to include sections for
both servers and workstations
* added the outcomes from the tests for servers w/o anacron, showing
that what the run-times are for cron.daily, cron.weekly, and
cron.monthly jobs
* changed current admonition from a "note" to a "tip" and referenced the
"tip" in the "For Workstations" descriptions of the cron "dot"
directories
* added a "note" admonition too to explain that there was no hard and
fast rule for server administrators not to use this method if they found
a use for it.
* Added clarification to the "Create Your Own `cron`" section
referencing the two different "dot" directory entries.
* Reworded the @options section introductory paragraph to make it
clearer
* Reworded the "note" admonition to make it more concise
* Added links to the two additional crontab related procedures in the
documentation in the conclusion section.
* Added an optional use of the `man crontab` to discover more command
options.