Suggested changes to chapter 3 - commands
* contributor had added in emojis in place of the admonitions, but this
removed some of the color highlighting that drew attention to the
admonition type in the original. Reverted to admonitions.
* contributor added a couple of options including -s to the `cp` command
and -u to the `sort` ... while these *may* be covered in the advanced
commands section (couldn't remember and didn't check to see) it didn't
seem like they were particularly problematic.
* added the author and contributors meta tags as these were not in the
original and added an update, which shows the date of the most recent
update.