add network.it.md
In this file I made some changes to the admonitions because I realized that translating the tags of the admonitions these were translated into Italian but all became of type note (blue color) !!!
So inspired by what was found in netwok.md:

!!! Notes "Memory aid"

I changed the admonitions tags like this:

!!! Warning Attenzione
!!! Tip Consiglio

This from my visualization tests produces both the right color and translation.
Can you please check it out?

I think it is important that the admonitions have the right color to indicate the level of attention.
Thank you