Edit PR 392
* reverted the "-c" option added in the English version of the
isntallation.md document to the wget command. This really shouldn't be
in the Chinese translations either, but I've left it alone for now. It
should be used to continue a stalled or incomplete download, but it is
used in the context of a fresh instance, which is not correct.
* Remove tip for Xshell in both migrate2rocky.md and
migrate2rocky.zh.md. This is a proprietary shell, not available from any
repository in Rocky. Instead of adding a tip corresponding to this
shell, which would only affect a handful of people, I encourage the
contributor to write a guide on installing and using Xshell instead.
* Added in spaces after periods where content was moved. In most cases
I've left the content alone, BUT, I've added back in some of the
author's own words in the English version of the migrate2rocky.md
script. I've said before that we like to maintain the voice of the
original author. A contributor may not like that voice, and they are
free (in this case) to modify them in the translation to their language,
but the original voice and generally the original words should remain.
* meta content was added where missing