Issue: Chinese language showing 404 error
* This was an issue with the .pages file. It needed to be changed so
that it would "see" all of the language files that exist in the
release_notes directory.
* Updated .pages file removing the `- Change Log: 8-changelog.md` line
and changing it to `- ... | 8-changlog*.md` which will cover all other
languages that the changelog might be translated into.
* Updated the 8-changelog.md and 8-changelog.zh.md documents to:
  - fix the `title:` meta so that the title is simply "Change Log"
- fix the `update:` meta so that it uses the 3-letter month format.
This has been done because of confusion being caused by dates entered
using international formats or US formats. Replacing the month with the
3-letter representation means that the update date can be entered by
anyone and is confusing to no one. Examples: Dec-07-2021 or 07-Dec-2021