Multi doc fixes
* Nextcloud: change "test editor" to "text editor" and add tags
* DokuWiki: add meta, fix more than 1 "#" heading issue, fix typo "Go
ahead an link..." to "Go ahead and link..."
* MATE: fix conclusion paragraph which did not make sense. Added known
meta and tags.
* DNS Bind: Fix "each number separated by a comma" to correct "each
number separated by a period". Fix "down one so that is" to "down one so
that it is". Add meta and tags.
* Micro: Fix "However, if you want to check the it out..." to "However,
if you want to check it out..." Add tags to meta.
* VSFTPD: Fix unclear sentence to two sentences: "We need to create a
file to read the virtual users and passwords from. This will create the
database." Add tags.
* Networking Configuration: Fix spelling of "automatically". Add in code
blocks. Fix code blocks. Add an appropriate admonition in place of bold
text. Reworded one sentence to: "The following shows an example for the
default configuration of a server:"
* Let's Encrypt: Fix multiple "#" headings. Fix "this deals with the
next line that regarding..." to "this deals with the next line
regarding..." Add tags.
* SSH Public and Private Keys: Fix multiple "#" headings. Fix spelling
mistake. Add admonition in place of bold text. Minor change to wording
in Introduction. Add meta and tags.
* Generating SSL Keys: Fix multiple "#" headings. Change: "causes higher
CPU load..." to "and causes higher CPU load". Fix formatting on
everything below: "This opens up a dialog..." Add meta and tags.
* Nginx-Mainline: Reword the paragraph that started with "Now to open up
the necessary ports..." Fix the line that read "you should make you can
run" to "you should run..." Fix "The time is coming when browsers
maight..." to "The time is coming when browsers might..." and remove
"even" in the same line. Add tags to meta.