Books section - Organizational rework
* This section's .pages files had gotten out of sync. First, the main
.pages file in the main directory did not have anything added to it
since the system administrator's guide was added. Instead of fixing
tiltes in that section, the titles had been fixed within the individual
directories in their own .pages files, which while works, adds to the
overall bloat of the .pages system. The only directory which actually
needed its own .pages file, was the newest one (learning bash), which
has the appendix added on for script examples.
* fixed all of this by:
- amending the main .pages file to add the titles for the sections and
the navigation to the correct directories.
- removed the individual directory .pages files EXCEPT for the one for
learning bash, which was modified to clean up the listing.
* tested all of this in my local mkdocs instance to make sure that it
looks like it is supposed to AND works