Force retranslation
Issue: translator noticed some old translated files with bad meta at the
top of the file. This doesn't happen with "new" translations. This
commit is designed to force a retranslation of those files to see if
this fixes the meta.

* minor changes made to rsnapshot_backup, cloud_server_using_nextcloud,
dokuwiki_server, and lxd_server

There are more of these files out there, so if this fixes things, then
we will address the rest in succession.