Formatting - `installation.md`
* To be clear, there is nothing wrong with the way the document was
formatted to begin with, BUT crowdin was seeing things like italics and
bold face as <0>item</0> instead of the appropriate tag. This is a
theory that the itallics (_) which can also be displayed as a single (*)
and is most commonly done that way throughout the rest of the
documentation, is being misinterpretted. This is true for the bold
(double _) as well, which can be (double *) instead.