Changes to installation documents for 8.6
* when 9.0 came out, the version of the installation guide for 8.6 was
moved to `8.6_installation.md` which cannot be translated. Moved this
file to `8_6_installation.md`
* this file _had_ been called `installation.md` but this is what the 9.0
installation guide became, and because of that, all of the translated
files beneath it (`installation.fr.md`, `installation.it.md` etc.), no
longer aligned with the correct version of the source document.
* moved each translated version of the 8.6 documents to
`8_6_installation.[language].md`
* changed the title of the `8_6_installation.md` to "Installing Rocky
Linux 8.6"

At some point the Installation documents for 8.6 and all of the language
versions will need to be moved into a separate folder for rocky specific
files and out of the documentation root where they now reside, but that
is for another PR.