# Internal link fixes: (#1494)
Because of the change in location for the "Local Docs" routines, some translated files are showing the old internal links, because the new link locations have not been approved in the translation interface. This is an attempt to fix those outside of Crowdin. Since it is just links, I'm hoping it will work as expected. There was also a link to an image file that had the "images" portion of the link translated, which should not have been. I've attempted to fix this as well.