Syntax of the undo command was incorrect (#2040)
* When markdown errors were recently corrected, the undo command was improperly formatted as
```
Undo the last action:

++u++

Undo the actions on the current line:

++U++
```
However neither line is correct. Markdown keyboard commands always assume a capital letter, and must be entered in lower case or in quotes. To get the desired lower-case "u" for the first and capital U for the second, it must be entered as:

```
Undo the last action:

++"u"++

Undo the actions on the current line:

++u++
```

The first will now render in lower-case as expected and the second will render a capital "U."