Update lab6-the_file_system.md (#2082)
Moved the "To remove a LV, VG and PV" section after the "To create a logical volume" section within Exercise 1, this seemingly being placed incorrectly due to the "scratch" variables appearing before they were defined. 

Added a step 4 to the "To assign a physical volume to a volume group" section, undoing the previous action, which will help the user follow into the next section.

Added a number 3 after "## Exercise".