Fix typos in package list (installing_nvidia_gpu_drivers.md) (#2408)
Fixed typos in list of necessary packages for building and installing kernel modules:

* libglvnd-devel instead of libglv-devel (does not exist)
* pkgconf instead of pkgconfig (or should it be pkg-config?)

The first item causes an error "Error: Unable to find a match: libglv-devel". Given that the list includes "libglvnd-opengl libglvnd-glx" I assume that "libglvnd-devel" is the right replacement.

It seems the second item is still correctly resolved (maybe there is an alias name for this package?). However, "pkgconfig" seems not to be the correct name.