docs: Update remi-release from 9 to 10, and php-fpm version to latest. (#2935)
* docs: Update remi-release from 9 to 10, and php-fpm version to latest.

Updated this for Rocky linux 10, however the current text "(note: if you are running Rocky Linux 8.x or 10.x, substitute in 8 or 10 next to the "release-" below):" may be sufficient already. I updated php-fpm aswell, as latest versions are in my opinion better, in order to avoid using potentially vulnerability-prone versions.

* Update caddy.md

---------

Co-authored-by: wale soyinka <wsoyinka@gmail.com>