Remove 'sudo' from the newgrp command example
Removed `sudo` that was in front of `newgrp libvirt`. `sudo` is not needed here as a regular user. Running `sudo` causes the user to be elevated into a `root` session with the command prompt and then the user has to type `exit` to get out of the `root` session. Not a big deal, however it helps to avoid user confusion.